+++ crates/bevy_render/src/renderer/render_resource_context.rs
@@ -38,14 +38,14 @@ pub trait RenderResourceContext: Downcast + Send + Sync + 'static {
         &self,
         handle: HandleUntyped,
         resource: RenderResourceId,
-        index: usize,
+        index: u64,
     );
     fn get_asset_resource_untyped(
         &self,
         handle: HandleUntyped,
-        index: usize,
+        index: u64,
     ) -> Option<RenderResourceId>;
-    fn remove_asset_resource_untyped(&self, handle: HandleUntyped, index: usize);
+    fn remove_asset_resource_untyped(&self, handle: HandleUntyped, index: u64);
     fn create_render_pipeline(
         &self,
         pipeline_handle: Handle<PipelineDescriptor>,
@@ -63,21 +63,21 @@ pub trait RenderResourceContext: Downcast + Send + Sync + 'static {
 }
 
 impl dyn RenderResourceContext {
-    pub fn set_asset_resource<T>(&self, handle: Handle<T>, resource: RenderResourceId, index: usize)
+    pub fn set_asset_resource<T>(&self, handle: Handle<T>, resource: RenderResourceId, index: u64)
     where
         T: 'static,
     {
         self.set_asset_resource_untyped(handle.into(), resource, index);
     }
 
-    pub fn get_asset_resource<T>(&self, handle: Handle<T>, index: usize) -> Option<RenderResourceId>
+    pub fn get_asset_resource<T>(&self, handle: Handle<T>, index: u64) -> Option<RenderResourceId>
     where
         T: 'static,
     {
         self.get_asset_resource_untyped(handle.into(), index)
     }
 
-    pub fn remove_asset_resource<T>(&self, handle: Handle<T>, index: usize)
+    pub fn remove_asset_resource<T>(&self, handle: Handle<T>, index: u64)
     where
         T: 'static,
     {
